assembly: fix case of unbounded growth of residual cache
Previously, assembly was caching residual values in vectors for both
time and non-time based kernels (KT_TIME and KT_NOTIME) regardless of whether
or not both kernel types were in use in the (nonlinear) system.  However, when
the caches were applied to actual residual vectors, they were conditionally
applied based on whether or not each kernel type was in use in the system.
Because the cache clearing/reset only occurred in conjunction with application
of the cache to actual residual vectors, simulations that used the cache but
didn't have one or more of the kernel types would have cache vectors for the
unused types grow unbounded. This change modifies the cache to only store
residual entries for kernel types that are in use.  It also clears the cache
vectors for all kernel types with the new addCachedResiduals function
regardless of which types are in use (just for safety).

This was discovered after noticing unbounded heap usage growth for
long-running rattlesnake simulations while doing memory profiling.

Fixes #9202