Move the try/exception to include unknown prereqs
We were building a dictionary of tests and then comparing those against user-generated
list of prereqs. Typos ensued, and we ran into a dictionary KeyError exception.

Include the dictionary comparison in a try/except to catch this error, and handle it
accordingly.

Add a unit test to test for this issue.

Refs #9224