Lock thread pool if tester has dependents
If the test in question has a dependent, we need to lock the thread pool while
we process this test's results. The dependent test can launch _as soon as_ this
test sets a passing status.

We should look into why we need to thread lock in this case, because according
to the tester, the last thing we set is a status. So the results file in question
should have remain unmolested until we set a passing status. But that does not
appear to be the case.

For now, if a test has dependents, lock the thread until the process has completed.

Refs #9280