Add --recoversuffix to run_tests
This makes it easier to do recovery via ASCII, which in turn is
turning out to be necessary while debugging #8890.

I've verified that the failure case in that issue is behaving (and
eventually failing) identically no matter whether we do binary or
ASCII recovery.  At some point I'll run the whole test suite and see
if any changes trigger elsewhere.