test: fix spaces in FunctionName typed input param
When an InputParameters value was templated with the c++ FunctionName
type, the getpot parser would only return the characters up to the first
whitespace character - even if the value was quoted properly. e.g.:

    paramWithFunctionNameType = "x*x + y*y"

would be truncated to x*x by the getpot parser (at least with the way we
were using it). There is one instance of this in the moose test
postprocessors/element_l2_difference.test that was fixed. That
test had a corresponding bad gold file that was regolded as well.

fixes #9623