Skip restart partitioning and renumbering
This fixes occasional bugs when parallel restart meshes are hit by
non-idempotent repartitioning.  In particular it fixes the
restart/new_dt.test_restart test with --distributed-mesh --recover
settings.  That test is an example of #5877, mixing recovery with
restarts, and it now works on 1 through 24 processors, so there's a
chance that that capability is working now.

This fixes some but not all of the potential DistributedMesh restart
problems discussed in #8410.  I'll open a new ticket for the remaining
issue.