Smeared Cracking model name change ref #9227
Since ComputeElasticSmearedCrackingStress no longer requires the underlying
model to be elastic, rename it to ComputeSmearedCrackingStress. Since this
is a new model that is not being widely used yet, I'm not deprecating the
old name.

Also, change 'cracking_stress' to be a coupled variable rather than a function.
If a single value is prescribed, the interface is the same. This allows
this model to be easily used with a randomly-generated field variable that
defines the strength.