Remove unsafe PenetrationInfo copy constructor
I don't get it.  This is clearly a double-delete bug.  The comment in
the implementation *acknowledges* that it's a double-delete bug.  The
only thing saving us from ridiculously-hard-to-debug memory corruption
is the fact that nothing (in framework/, test/, or modules/, at least)
actually uses this time bomb of a copy constructor.